+++ lib/dart_style.dart
@@ -6,3 +6,4 @@ library dart_style;
 
 export 'src/dart_formatter.dart';
 export 'src/formatter_exception.dart';
+export 'src/source_code.dart';